Consolidate event-manager-related classes into utils/events
